+++ jib-core/src/test/java/com/google/cloud/tools/jib/registry/ManifestPusherTest.java
@@ -123,7 +123,7 @@ public class ManifestPusherTest {
       Assert.assertThat(
           ex.getMessage(),
           CoreMatchers.containsString(
-              "Registry does not support Image Manifest Version 2, Schema 2"));
+              "Registry may not support Image Manifest Version 2, Schema 2"));
 
     } catch (HttpResponseException ex) {
       Assert.fail("should have been a RegistryErrorException");